pyproject.toml
[project]
name = "dev-project"
version = "0.1.0"
description = ""
authors = [
    {name = "SGreshilov",email = "workgresh@gmail.com"}
]
readme = "README.md"
requires-python = "~3.11"
dependencies = [
    "django (>=4.1,<4.2)",
    "django-admin-list-filter-dropdown (>=1.0.3,<2.0.0)",
    "djangorestframework (==3.15.1)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
